Revert "UltimosCambios 07/04"
This reverts commit a92f00fce514a134473b3d218392cd0b766841c0.

+++ macowins/src/main/java/Grupo4/Negocio.java
@@ -12,7 +12,7 @@ public class Negocio
 	List <Venta> ventas=new ArrayList<>();
 	
 	
-    public double precioFinal(Prenda unaPrenda){
+    public float precioFinal(Prenda unaPrenda){
     		return unaPrenda.getPrecioFinalPrenda();
     }
 	public void realizarVenta(Prenda unaPrenda,int cantidad){
@@ -23,11 +23,11 @@ public class Negocio
     private void registrarVenta(Venta unaVenta){
     	ventas.add(unaVenta);
     }
-    public double gananciasDelDia(String fecha){ //Formato Fecha dd/m/aaaa
+    public float gananciasDelDia(String fecha){ //Formato Fecha dd/m/aaaa
     	List <Venta> ventasDelDia=new ArrayList<>();
-    	double ganancia;
+    	float ganancia;
     	ventasDelDia=ventas.stream().filter(unaVenta->unaVenta.getFecha().equals(fecha)).collect(Collectors.toList());
-    	ganancia= ventasDelDia.stream().mapToDouble(unaVenta->(unaVenta.getPrecioFinal())).sum();
+    	ganancia=(float) ventasDelDia.stream().mapToDouble(unaVenta->(unaVenta.getPrecioFinal())).sum();
     	return ganancia;
     }
     private String calcularFecha(){
