+++ macowins/src/main/java/Grupo4/Pantalon.java
@@ -1,12 +1,12 @@
 package Grupo4;
 
 public class Pantalon implements Tipo {
-	private double cantTela;
+	private float cantTela;
 	
-	public Pantalon(double cantTela) {
+	public Pantalon(float cantTela) {
 		this.cantTela = cantTela;
 	}
-	public double precioBase(){
+	public float precioBase(){
 		return 250+cantTela;
 	}
 }
